requirements.txt
requests
flask
flask-cors
supabase
python-dotenv
bs4